dynamic update property form, add custom tree form
