+++ src/compas_view2/app/selector.py
@@ -2,7 +2,7 @@ import time
 from random import randint
 import numpy as np
 
-from .worker import Worker
+from .worker import Ticker
 
 
 class Selector:
@@ -109,34 +109,32 @@ class Selector:
         None
 
         """
-        self.stop_monitor_instance_map = False
+        def select():
+            if self.instance_map is not None:
+                instance_map = self.instance_map
+                if self.select_from == "pixel":
+                    # Pick an object from mouse pixel
+                    x = self.app.view.mouse.last_pos.x()
+                    y = self.app.view.mouse.last_pos.y()
+                    self.select_one_from_instance_map(x, y, instance_map)
+                if self.select_from == "box":
+                    # Pick objects from box selection
+                    self.select_all_from_instance_map(instance_map)
+                    self.select_from = "pixel"
+                self.app.view.update()
+                self.instance_map = None
 
-        def monitor_loop():
-            while not self.stop_monitor_instance_map:
-                time.sleep(0.02)
-                if self.instance_map is not None:
-                    instance_map = self.instance_map
-                    if self.select_from == "pixel":
-                        # Pick an object from mouse pixel
-                        x = self.app.view.mouse.last_pos.x()
-                        y = self.app.view.mouse.last_pos.y()
-                        self.select_one_from_instance_map(x, y, instance_map)
-                    if self.select_from == "box":
-                        # Pick objects from box selection
-                        self.select_all_from_instance_map(instance_map)
-                        self.select_from = "pixel"
-                    self.app.view.update()
-                    self.instance_map = None
-
-        # Stop the monitor loop when the app is being closed
-        def stop():
-            self.stop_monitor_instance_map = True
-        self.app._app.aboutToQuit.connect(stop)
+                if self.app.dock_slots["propertyform"] is not None and self.selected:
+                    self.app.dock_slots["propertyform"].set_object(self.selected[0])
+                if self.app.dock_slots["sceneform"] is not None:
+                    self.app.dock_slots["sceneform"].select(self.selected)
 
         # Start monitor loop in a separate worker thread
-        worker = Worker(monitor_loop)
-        worker.no_signals = True  # Prevent signal emit error when closing the app
-        Worker.pool.start(worker)
+        ticker = Ticker(interval=0.02)
+        ticker.signals.tick.connect(select)
+        Ticker.pool.start(ticker)
+
+        self.app._app.aboutToQuit.connect(ticker.stop)
 
     def get_rgb_key(self):
         """
