+++ src/compas_view2/app/worker.py
@@ -1,5 +1,6 @@
 import sys
 import traceback
+import time
 from qtpy.QtCore import QObject
 from qtpy.QtCore import QRunnable
 from qtpy.QtCore import QThreadPool
@@ -13,6 +14,7 @@ class WorkerSignals(QObject):
     finished = Signal()
     error = Signal(tuple)
     result = Signal(object)
+    tick = Signal()
     progress = Signal(object)
 
 
@@ -51,3 +53,24 @@ class Worker(QRunnable):
                 self.signals.result.emit(result)
             finally:
                 self.signals.finished.emit()  # Done
+
+class Ticker(QRunnable):
+
+    pool = QThreadPool()
+
+    def __init__(self, interval=1):
+        super(Ticker, self).__init__()
+        self.interval = interval
+        self.signals = WorkerSignals()
+
+    @Slot()
+    def run(self):
+        """Execute the worker function, send on signals on different scenarios
+        """
+        self.running = True
+        while self.running:
+            self.signals.tick.emit()
+            time.sleep(self.interval)
+    
+    def stop(self):
+        self.running = False
