+++ src/compas_view2/forms/__init__.py
@@ -1,8 +1,8 @@
 from .form import Form  # noqa: F401
-from .editform import EditForm  # noqa: F401
+from .propertyform import PropertyForm  # noqa: F401
 from .addform import AddForm  # noqa: F401
-from .point import PointForm  # noqa: F401
-from .line import LineForm  # noqa: F401
-from .sphere import SphereForm  # noqa: F401
-from .torus import TorusForm  # noqa: F401
-from .pointedit import PointEditForm  # noqa: F401
+# from .point import PointForm  # noqa: F401
+# from .line import LineForm  # noqa: F401
+# from .sphere import SphereForm  # noqa: F401
+# from .torus import TorusForm  # noqa: F401
+# from .pointedit import PointEditForm  # noqa: F401
